* Fixed serialbee config parsing.

+++ modules/serialbee/serialbeefactory.go
@@ -33,7 +33,7 @@ type SerialBeeFactory struct {
 func (factory *SerialBeeFactory) New(name, description string, options modules.BeeOptions) modules.ModuleInterface {
 	bee := SerialBee{
 		device:      options.GetValue("device").(string),
-		baudrate:        options.GetValue("baudrate").(int),
+		baudrate:        int(options.GetValue("baudrate").(float64)),
 	}
 
 	bee.Module = modules.Module{name, factory.Name(), description}
